@@ -1,3 +1,3 @@
 module EdgeUpdater
 
-go 1.15
+go 1.19
